Add routing and fix jest path
