+++ package.json
@@ -18,7 +18,8 @@
     "jest-environment-jsdom": "^29.7.0",
     "monaco-editor": "^0.49.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.4.5",
